% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{ClassifyByProfileVector.SMR}
\alias{ClassifyByProfileVector.SMR}
\title{Classification using a profile vector}
\usage{
\method{ClassifyByProfileVector}{SMR}(x, tagType, profileVec, nTopNNs,
  voting = FALSE)
}
\arguments{
\item{x}{A recommender object.}

\item{tagType}{Tag type for which the classification is done.}

\item{profileVec}{Is a sparse matrix with 1 row (a row from a sparse matrix).}

\item{nTopNNs}{Number of top nearest neighbors to be used in to derive the classification.}

\item{voting}{Should simple voting be used or a weighted sum?}
}
\value{
A list of named scores.
}
\description{
Specialization of \code{\link{ClassifyByProfileVector}}
for sparse matrix recommender objects.
Classify a profile vector into the column names of a tag type sub-matrix.
}
