% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{ClassifyByProfileVector.TSCorrSMR}
\alias{ClassifyByProfileVector.TSCorrSMR}
\title{Classification with time series recommender}
\usage{
\method{ClassifyByProfileVector}{TSCorrSMR}(x, profileVec, nTopNNs,
  voting = FALSE)
}
\arguments{
\item{x}{A time series recommender object ("TSCorrSMR").}

\item{profileVec}{A profile vector.}

\item{nTopNNs}{Number of top nearest neighbors to be used.}

\item{voting}{Should voting method to be used?}
}
\value{
A data frame with columns "Score", "Label".
}
\description{
Specialization of \code{ClassifyByProfileVector} for time series recommender objects.
}
\details{
The two methods of making the classifications from the obtained recommendations 
are: "voting" and "scoring".
}
