% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{ConsumptionProfile}
\alias{ConsumptionProfile}
\title{Calculation of a consumption profile}
\usage{
ConsumptionProfile(x, historyItems, historyRatings, allColumns = FALSE,
  ...)
}
\arguments{
\item{x}{a recommender object}

\item{historyItems}{a list of history items (indices or ID's)}

\item{historyRatings}{a list of history ratings}

\item{allColumns}{a logical are all columns of the results returned or not}
}
\value{
A data frame with the first columns being "Score" and "Tag".
}
\description{
Generic function for calculating a consumption profile.
}
