% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfMushroom}
\alias{dfMushroom}
\title{Mushroom edibility}
\format{A data frame with 8124 rows and 24 columns.}
\source{
\url{https://datahub.io/machine-learning/mushroom}
}
\usage{
dfMushroom
}
\description{
This data set consists of 8124 records of the physical 
charachteristics of gilled mushrooms in the Agaricus and Lepiota family,
along with their edibility. 
Each row represents one mushroom.
}
\keyword{datasets}
