% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{MakeTimeSeriesSearchVectors}
\alias{MakeTimeSeriesSearchVectors}
\title{Search time series creation}
\usage{
MakeTimeSeriesSearchVectors(tsMat)
}
\arguments{
\item{tsMat}{A sparse matrix with rows corresponding to time series.}
}
\description{
Creates a list of search vectors (time series) for a given time series matrix.
}
\details{
This "helper" function can be used for finding trends of the signals.
}
