% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{predict.SMR}
\alias{predict.SMR}
\title{Classification with a sparse matrix recommender object}
\usage{
\method{predict}{SMR}(smr, data, type = "decision", normalizedQ = TRUE, ...)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{data}{A matrix or a data frame.}

\item{type}{What kind of result to be returned?
The value 'raw' returns a matrix, 'decision' a vector of labels.}

\item{normalizedQ}{Should the results be normalized or not if type equals 'raw'?}

\item{...}{Parameters \code{\link{SMRClassifyByProfileVector}}.}
}
\value{
If \code{type="decision"} -- a vector of scored class labels.
If \code{type="raw"} or \code{type="scores"} -- a contingency matrix of labels and scores.
}
\description{
Classify a data frame or matrix using a sparse matrix recommender.
}
\details{
The sparse matrix recommender \code{smr} can have additional parameters
tucked-in to \code{smr['ClassifierParameters']}.
The list \code{smr['ClassifierParameters']} can have elements: 
"tagType", "nTopNNs", "voting", "dropZeroScoredLabels". (Same as \code{...}.)
}
