% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{Recommendations.CompositeRecommender}
\alias{Recommendations.CompositeRecommender}
\title{Calculate recommendations with a composite recommender}
\usage{
\method{Recommendations}{CompositeRecommender}(x, historyItems,
  historyRatings, nrecs, removeHistory = TRUE,
  normalizationType = NULL, mergeFunction = NULL, ...)
}
\arguments{
\item{x}{A recommender object.}

\item{historyItems}{A list of history items.}

\item{historyRatings}{A list of history ratings.}

\item{nrecs}{Number of required recommendations.}

\item{removeHistory}{Should the history be removed or not?}

\item{normalizationType}{Normalization type, one of NULL, 'none', 'max', 'rank', 'quantileIntervals',
or 'shiftAndRescale' (same as NULL).}

\item{mergeFunction}{A function to merge the recommendations lists, a function that can be applied to
a vector of scores corresponding to an item.}
}
\value{
A data frame with the columns columns "Score", "Item".
}
\description{
Calculate recommendations over a composite recommender object.
}
\details{
If the argument \code{normalizationType} is NULL, then the object's element 'NormalizationType' is used.
If that is NULL too, then 'shiftAndRescale' is used. Examples of values of \code{mergeFunction} are 'sum',
'max', 'mean', 'median', 'length'. If \code{mergeFunction} is NULL, then the object's element 'MergeFunction'
is used. If that is NULL too, then sum is used.
}
