% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{Recommendations}
\alias{Recommendations}
\title{Compute recommendations}
\usage{
Recommendations(
  x,
  historyItems,
  historyRatings,
  nrecs,
  removeHistory = TRUE,
  ...
)
}
\arguments{
\item{x}{A recommender object.}

\item{historyItems}{A list of history items (indices or ID's).}

\item{historyRatings}{A list of history ratings.}

\item{nrecs}{Number of required recommendations.}

\item{removeHistory}{Should the history be dropped or not?}
}
\value{
A data frame with the columns c("Score", "Item").
}
\description{
The generic function for calculating recommendations by history.
}
\seealso{
Other Recommendations calculation functions: 
\code{\link{Recommendations.SMR}()}
}
\concept{Recommendations calculation functions}
