% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{Recommendations.SMR}
\alias{Recommendations.SMR}
\title{Specialization of Recommendations for SMR objects.}
\usage{
\method{Recommendations}{SMR}(x, historyItems, historyRatings, nrecs,
  removeHistory = TRUE, ...)
}
\arguments{
\item{x}{A recommender object.}

\item{historyItems}{A list of history items (indices or ID's).}

\item{historyRatings}{A list of history ratings.}

\item{nrecs}{Number of required recommendations.}

\item{removeHistory}{Should the history be dropped or not?}
}
\value{
The result is a data frame with the columns c("Score", "Item").
}
\description{
Specialization of Recommendations for SMR objects.
}
\seealso{
Other Recommendations calculations: \code{\link{Recommendations}}
}
\concept{Recommendations calculations}
