% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{Recommendations.TSCorrSMR}
\alias{Recommendations.TSCorrSMR}
\title{Time series recommendations calculation}
\usage{
\method{Recommendations}{TSCorrSMR}(
  x,
  historyItems,
  historyRatings,
  nrecs,
  removeHistory = TRUE,
  ...
)
}
\arguments{
\item{x}{A recommender object.}

\item{historyItems}{A list of history items.}

\item{historyRatings}{A list of history ratings.}

\item{nrecs}{Number of required recommendations.}

\item{removeHistory}{Should the history be removed or not?}
}
\value{
A data frame with columns "Score", "Item".
}
\description{
Specialization of Recommendations for TSCorrSMR objects.
}
