% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{RecommendationsByProfile}
\alias{RecommendationsByProfile}
\title{Compute recommendations by profile.}
\usage{
RecommendationsByProfile(x, profileTags, profileTagScores, nrecs, ...)
}
\arguments{
\item{x}{A recommender object.}

\item{profileTags}{A list of profile tags.}

\item{profileTagScores}{A list of scores corresponding to the profile tags.}

\item{nrecs}{Number of required recommendations.}
}
\value{
A data frame with the columns columns c("Score", "Item")
}
\description{
The generic function for calculating recommendations by profile.
}
