% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{RecommenderItems}
\alias{RecommenderItems}
\title{Recommender items}
\usage{
RecommenderItems(recommender)
}
\arguments{
\item{recommender}{A recommender object.}
}
\value{
A list of item ID's (strings).
}
\description{
Recommender items
}
