% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{RecommenderTags}
\alias{RecommenderTags}
\title{Recommender tags}
\usage{
RecommenderTags(recommender)
}
\arguments{
\item{recommender}{A recommender object.}
}
\value{
A list of tags.
}
\description{
Recommender tags
}
