% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRApplyTagTypeWeights}
\alias{SMRApplyTagTypeWeights}
\title{Change tag type weights}
\usage{
SMRApplyTagTypeWeights(smr, weights)
}
\arguments{
\item{smr}{A sparse matrix recommender object (list with named elements).}

\item{weights}{A list/vector of weights to be applied.}
}
\value{
Sparse Matrix.
}
\description{
Changes the weights of tag types of a sparse matrix recommender object.
}
\details{
If \code{weights} does not have names it is going to replicated to match 
the length of \code{smr$TagTypes}.
If \code{weights} has names the missing tag types are (if any) 
are set to have weight 1.
Note that the result of this function is a sparse matrix not an SMR object.
(The function name probably have to be changed for consistency: 
tag types should have "significance factors" not "weights".)
}
