% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRApplyTagWeights}
\alias{SMRApplyTagWeights}
\title{Apply tag weights}
\usage{
SMRApplyTagWeights(smr, weights)
}
\arguments{
\item{smr}{A sparse matrix recommender object (list with named elements).}

\item{weights}{A list of weights to be applied.}
}
\value{
Sparse matrix.
}
\description{
Changes the weights of the tags of a sparse matrix recommender object
}
\details{
Currently there is some mixture (confusion) between "significance factors"
of tag types or tags and "weights" of matrix entries.
}
