% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRApplyTermWeightFunctions}
\alias{SMRApplyTermWeightFunctions}
\title{Document-term matrix term re-weighting.}
\usage{
SMRApplyTermWeightFunctions(docTermMat, globalWeightFunction = NULL,
  localWeightFunction = NULL, normalizerFunction = NULL)
}
\arguments{
\item{docTermMat}{A document-term sparse matrix (dgCMatrix).}

\item{globalWeightFunction}{A global weight function ID (a string, one of "IDF", "GFIDF", "Normal", "None").}

\item{localWeightFunction}{A global weight function ID (a string, one of "Binary", "TermFrequency", "Log", "None").}

\item{normalizerFunction}{A normalization weight function ID (a string, one of "Cosine", "Sum", "None"),}
}
\value{
Sparse matrix
}
\description{
Applies the global weight functions like Inverse Document Frequency (IDF) to the entries of a sparse matrix.
}
\details{
The implemented global weight function ID's are "IDF", "GFIDF", "Normal", "Binary", "Entropy", "Sum", "None".
The implemented local weight function ID's are "Binary", "TermFrequency", "Log", "Logarithmic", "None".
The implemented normalization function ID's are "Cosine", "Sum", "None".
}
\seealso{
Other Document-term matrix functions: \code{\link{SMRMakeColumnStochastic}},
  \code{\link{SMRMakeDocumentTermMatrix}},
  \code{\link{SMRToBagOfWords}}
}
\concept{Document-term matrix functions}
