% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCategorizeToIntervals}
\alias{SMRCategorizeToIntervals}
\title{Categorize to intervals.}
\usage{
SMRCategorizeToIntervals(
  vec,
  breaks = NULL,
  probs = seq(0, 1, 0.1),
  intervalNamesQ = FALSE
)
}
\arguments{
\item{vec}{A numerical vector to be categorized.}

\item{breaks}{A the breaks to be used.
If NULL \code{quantile} is used.}

\item{probs}{A numerical vector to be given to 
the argument \code{probs} of \code{\link{quantile}}.}

\item{intervalNamesQ}{Should the intervals be represented with integers or with character names?}
}
\value{
A character vector.
}
\description{
Categorize to intervals.
}
