% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRClassifyByProfileVector}
\alias{SMRClassifyByProfileVector}
\title{Classification with a profile vector}
\usage{
SMRClassifyByProfileVector(smr, tagType, profileVec, nTopNNs,
  voting = FALSE, dropZeroScoredLabels = TRUE,
  maxNumberOfLabels = NULL, normalizeQ = TRUE)
}
\arguments{
\item{tagType}{Tag type for which the classification is done.}

\item{profileVec}{A sparse matrix with 1 row (a row from a sparse matrix).}

\item{nTopNNs}{Number of top nearest neighbors to be used in the derive the classification.}

\item{voting}{Should simple voting be used or a weighted sum?}

\item{maxNumberOfLabels}{The maximum number of labels to be returned; 
if NULL all found labels are returned.}

\item{normalizeQ}{Should the scores be normalized?
(By dividing by the maximum score.)}

\item{smar}{A sparse matrix recommender.}
}
\value{
A list of scored tags.
}
\description{
Classify a profile vector into the column names of a tag type sub-matrix.
}
