% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRClip}
\alias{SMRClip}
\title{Clip a sparse matrix.}
\usage{
SMRClip(smat, min = 0, max = 1, vmin = min, vmax = max)
}
\arguments{
\item{smat}{A sparse matrix.}

\item{min}{Lower boundary.}

\item{max}{Upper boundary.}

\item{vmin}{New value for values lesser than \code{min}.}

\item{vmax}{New value for values greater than \code{max}.}
}
\value{
Sparse matrix
}
\description{
Replaces non-zero entries with 1's.
}
\seealso{
Other Document-term matrix functions: 
\code{\link{SMRApplyTermWeightFunctions}()},
\code{\link{SMRGlobalTermFunctionWeights}()},
\code{\link{SMRMakeColumnStochastic}()},
\code{\link{SMRMakeDocumentTermMatrix}()},
\code{\link{SMRToBagOfWords}()},
\code{\link{SMRUnitize}()}
}
\concept{Document-term matrix functions}
