% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromLongForm}
\alias{SMRCreateFromLongForm}
\title{Creation of a SMR object from a long form data}
\usage{
SMRCreateFromLongForm(
  data,
  itemColumnName = "Item",
  tagTypeColumnName = "TagType",
  valueColumnName = "Value",
  weightColumnName = "Weight",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{itemColumnName}{Item (names or ID's) column.}

\item{tagTypeColumnName}{Tag type column.}

\item{valueColumnName}{(Tag) value column.}

\item{weightColumnName}{Weight column.}

\item{...}{Additional parameters for \code{\link{SMRCreateFromMatrices}}.}
}
\value{
SMR object.
}
\description{
Creates a Sparse Matrix Recommender object from long form data.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: 
\code{\link{SMRCreateFromMatrices}()},
\code{\link{SMRCreateFromSpecification}()},
\code{\link{SMRCreateItemTagMatrix}()},
\code{\link{SMRCreate}()},
\code{\link{SMREmptySpecification}()},
\code{\link{TSCorrSMRCreate}()}
}
\concept{Creation functions}
