% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromMatrices}
\alias{SMRCreateFromMatrices}
\title{Creation of a Sparse Matrix Recommender object with matrix.}
\usage{
SMRCreateFromMatrices(matrices, tagTypes, itemColumnName)
}
\arguments{
\item{matrices}{Matrices to be spliced into a metadata matrix.}

\item{tagTypes}{The name of the column containing the categorical tags.}

\item{itemColumnName}{The name of the column containing the unique items.}
}
\value{
SMR object
}
\description{
Creates a sparse matrix recommender from a list of matrices and a corresponding list of tag types.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: \code{\link{SMRCreateItemTagMatrix}},
  \code{\link{SMRCreate}}
}
\concept{Creation functions}
