% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromSpecification}
\alias{SMRCreateFromSpecification}
\title{Creation of a Sparse Matrix Recommender object from a specification.}
\usage{
SMRCreateFromSpecification(data, metaDataSpec, itemCol,
  .progress = "none", .verbose = FALSE)
}
\arguments{
\item{data}{A transactions-like data frame.}

\item{metaDataSpec}{A data frame with specifications of which columns of
\code{data} to be used and with what weight functions.}

\item{itemCol}{The name of the column containing the unique items.}
}
\value{
SMR object
}
\description{
Creates a sparse matrix recommender from transactions-like data
and a meta-data specification.
}
