% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateItemTagMatrix}
\alias{SMRCreateItemTagMatrix}
\title{Creation of item-tag matrix.}
\usage{
SMRCreateItemTagMatrix(dataRows, itemColumnName, tagType, sparse = TRUE)
}
\arguments{
\item{dataRows}{A data frame corresponding to a item consumption metadata table.}

\item{itemColumnName}{A name of the column of \code{dataRows} the values of which correspond to the rows of the returned matrix.}

\item{tagType}{A name of the column of \code{dataRows} the values of which correspond to the columns of the returned matrix.}

\item{sparse}{A logical, should the returned matrix be sparse or not.}
}
\value{
(Sparse) matrix
}
\description{
Convert to contingency matrix from item consumption "transactions" (e.g. instances of movie watching).
}
\seealso{
Other Creation functions: \code{\link{SMRCreateFromMatrices}},
  \code{\link{SMRCreate}}, \code{\link{TSCorrSMRCreate}}
}
\concept{Creation functions}
