% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRCrossTabulateTagTypes}
\alias{SMRCrossTabulateTagTypes}
\title{Cross tabulate tag types}
\usage{
SMRCrossTabulateTagTypes(
  smr,
  tagType1,
  tagType2,
  value = "none",
  normalizerFunction = NULL
)
}
\arguments{
\item{smr}{An SMR object.}

\item{tagType1}{Tag type corresponding to the rows of the cross tabulation.}

\item{tagType2}{Tag type corresponding to the common of the cross tabulation.}

\item{value}{Which value to use in the cross tabulation formula?
One of \code{c("none", "left", "right", "multiply")}.
If "none" then \code{xtabs( ~ tagType1 + tagType2, ...)} is used.}

\item{normalizerFunction}{A string specifying the normalization function 
for the result sparse matrix. 
If NULL then no normalization is done.}
}
\value{
A sparse matrix
}
\description{
Cross tabulates two specified tag types.
}
\details{
Uses \code{\link{xtabs}} and \code{\link{SMRApplyTermWeightFunctions}}.
}
