% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCurrentTagTypeSignificanceFactors}
\alias{SMRCurrentTagTypeSignificanceFactors}
\title{Get current tag type significance factors}
\usage{
SMRCurrentTagTypeSignificanceFactors(smr)
}
\arguments{
\item{smr}{A sparse matrix object.}
}
\value{
A list of named significance factors (numbers).
}
\description{
Finds the current significance factors in a SMR object.
}
