% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMREmptySparseMatrixRecommender}
\alias{SMREmptySparseMatrixRecommender}
\title{Empty sparse matrix recommender}
\usage{
SMREmptySparseMatrixRecommender()
}
\value{
A list
}
\description{
Makes an empty sparse matrix recommender object.
}
\details{
The result is a list that has element names a sparse matrix recommender.
}
