% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMREmptySpecification}
\alias{SMREmptySpecification}
\title{Empty specification}
\usage{
SMREmptySpecification(nrow = 1)
}
\arguments{
\item{nrow}{Number of rows.}
}
\value{
A data frame.
}
\description{
Creates an empty specification data frame for the creation of 
a sparse matrix recommender.
}
\details{
See \code{\link{SMRCreateFromSpecification}}.
}
\seealso{
Other Creation functions: 
\code{\link{SMRCreateFromLongForm}()},
\code{\link{SMRCreateFromMatrices}()},
\code{\link{SMRCreateFromSpecification}()},
\code{\link{SMRCreateItemTagMatrix}()},
\code{\link{SMRCreate}()},
\code{\link{TSCorrSMRCreate}()}
}
\concept{Creation functions}
