% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMREuclideanDistances}
\alias{SMREuclideanDistances}
\title{Euclidean distances from a reference vector per tag.}
\usage{
SMREuclideanDistances(smr, tagType = NULL)
}
\arguments{
\item{smr}{An SMR object.}

\item{tagType}{A string that is one of the tag types of \code{smr}.
If NULL distances for all tag types are computed.}
}
\value{
A data frame
}
\description{
For each tag of \code{tagType} this function does the following steps:
1) finds the sub-matrix of \code{smr$M} for which the corresponding tag column is not zero,
2) finds the mean vector of that sub-matrix,
3) computes the distances of the sub-matrix rows to the mean vector.
If \code{tagType = NULL} then for each element of \code{smr$TagTypes}
the corresponding sub-matrix distances-to-the-mean are computed.
}
