% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRExportToDirectory}
\alias{SMRExportToDirectory}
\title{Exports recommender to a directory}
\usage{
SMRExportToDirectory(smr, directoryPath, dataNameInfix = "", digits = NULL)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{directoryPath}{A path to a directory.}

\item{dataNameInfix}{A string designating the exported SMR object.}

\item{digits}{An integer for the number of digits to round to.
See \code{\link{round}}.
If NULL no rounding is done.
Only applies to the file with the name pattern "SMR-M01-from-*.csv".}
}
\value{
A logical.
}
\description{
Exports the recommender to specified directory.
}
\details{
The sparse matrix recommender \code{smr} is
exported using \code{\link{writeMM}}.
}
