% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRExportToFeatherFile}
\alias{SMRExportToFeatherFile}
\title{Export to feather format file}
\usage{
SMRExportToFeatherFile(smr, path)
}
\arguments{
\item{smr}{An SMR object.}

\item{path}{Path to feather file.}
}
\description{
Exports an SMR object to a feather format file.
}
\details{
Uses \code{\link{SMRMatricesToLongForm}} and \code{\link{feather::write_feather}}.
}
