% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRExtendProfile}
\alias{SMRExtendProfile}
\title{Extend profile}
\usage{
SMRExtendProfile(smr, profile, nrecs = NULL, normalizeQ = TRUE)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{profile}{A character vector of tags, a profile vector, or a profile data frame.}

\item{nrecs}{Number of recommendations.}

\item{normalizeQ}{Should the scores in result be normalized by the maximum score?}
}
\description{
Makes and extended profile of a given profile.
}
\details{
This function has the following steps. 
(1) Find recommendations for the given profile.
(2) Find the profile of the found recommendations.
(3) Optionally, normalize the scores.
}
