% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRGlobalTermFunctionWeights}
\alias{SMRGlobalTermFunctionWeights}
\title{Global weights for a document-term matrix.}
\usage{
SMRGlobalTermFunctionWeights(docTermMat, globalWeightFunction = "None")
}
\arguments{
\item{docTermMat}{A document-term sparse matrix (dgCMatrix).}

\item{globalWeightFunction}{A global weight function ID.}
}
\value{
Sparse matrix
}
\description{
Computes the global weights with a function specification.
}
\details{
The implemented global weight function ID's are "IDF", "GFIDF", "Normal", "Binary", "Entropy", "Sum", "None".
}
\seealso{
Other Document-term matrix functions: 
\code{\link{SMRApplyTermWeightFunctions}()},
\code{\link{SMRClip}()},
\code{\link{SMRMakeColumnStochastic}()},
\code{\link{SMRMakeDocumentTermMatrix}()},
\code{\link{SMRToBagOfWords}()},
\code{\link{SMRUnitize}()}
}
\concept{Document-term matrix functions}
