% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRHistoryProofs}
\alias{SMRHistoryProofs}
\title{History proofs}
\usage{
SMRHistoryProofs(smr, toBeLovedItem, history, normalizeScores = TRUE)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{toBeLovedItem}{An ID of a item or its index in \code{smr$M}.}

\item{history}{A data frame that is the customer purchasing history with columns "Score", <some-item-ID>.}

\item{normalizeScores}{Should the scores be normalized with \code{max(res$Score)}?}
}
\value{
A data frame with columns names "Score", <some-item-id>.
}
\description{
Find the items of the history that are the closest to a recommendation.
}
\seealso{
Other Proof functions: 
\code{\link{SMRMetadataProofs}()}
}
\concept{Proof functions}
