% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRImportFromDirectory}
\alias{SMRImportFromDirectory}
\title{Imports recommender from a directory}
\usage{
SMRImportFromDirectory(
  smr,
  directoryPath,
  dataNameInfix = "",
  itemColumnName = "ID"
)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{directoryPath}{A path to a directory.}

\item{dataNameInfix}{A string designating the exported SMR object.}

\item{itemColumnName}{A string to be set to results "ItemColumnName" element.}
}
\value{
A sparse matrix recommender object
}
\description{
Exports the recommender to specified directory.
}
\details{
The sparse matrix recommender \code{smr} is
imported using \code{\link{readMM}}.
}
