% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRImportFromFeatherFile}
\alias{SMRImportFromFeatherFile}
\title{Import from feather format file}
\usage{
SMRImportFromFeatherFile(path)
}
\arguments{
\item{path}{Path to feather file.}
}
\value{
An SMR object
}
\description{
Exports an SMR object to a feather format file.
}
\details{
Uses \code{\link{SMRCreateFromLongForm}} and \code{\link{feather::read_feather}}.
}
