% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRImposeColumnIDs}
\alias{SMRImposeColumnIDs}
\title{Impose column ID's.}
\usage{
SMRImposeColumnIDs(colIDs, smat)
}
\arguments{
\item{colIDs}{An array of column ID's.}

\item{smat}{A matrix with named columns.}
}
\value{
Matrix
}
\description{
Makes sure that the columns of a matrix are in 1-to-1 correspondence to an array of columns ID's.
}
\seealso{
Other Sparse matrix transformations: \code{\link{SMRImposeRowIDs}},
  \code{\link{SMRSparseMatrixToTriplets}},
  \code{\link{SMRTripletsToSparseMatrix}}
}
\concept{Sparse matrix transformations}
