% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRItemData}
\alias{SMRItemData}
\title{Interpret recommendations.}
\usage{
SMRItemData(smr, recs, tagTypes = NULL)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{recs}{A data frame of recommendations with column names "Score", "Index".}

\item{tagTypes}{A vector of tag types to use.}
}
\description{
Gives the interpretation of a data frame of recommendations with sparse matrix recommender object.
}
