% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRJoin}
\alias{SMRJoin}
\title{Join two sparse matrix recommender objects.}
\usage{
SMRJoin(smr1, smr2, colnamesPrefix1 = NULL, colnamesPrefix2 = NULL)
}
\arguments{
\item{smr1}{The first SMR object.}

\item{smr2}{The second SMR object.}

\item{colnamesPrefix1}{The prefix to be concatenated to the colnames of the first SMR object.}

\item{colnamesPrefix2}{The prefix to be concatenated to the colnames of the second SMR object.}
}
\value{
Sparse matrix recommender
}
\description{
Join two SMR objects.
}
\seealso{
Other SMR modification: \code{\link{SMRAnnexSubMatrix}},
  \code{\link{SMRRemoveTagTypes}},
  \code{\link{SMRRowBindMatrix}}, \code{\link{SMRRowBind}}
}
\concept{SMR modification}
