% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRMakeColumnStochastic}
\alias{SMRMakeColumnStochastic}
\title{Make a matrix column stochastic.}
\usage{
SMRMakeColumnStochastic(mat)
}
\arguments{
\item{mat}{A matrix (sparse or dense).}
}
\value{
Sparse matrix
}
\description{
Makes the matrix argument a column stochastic matrix (the sum of each column is 1).
}
\seealso{
Other Document-term matrix functions: \code{\link{SMRApplyTermWeightFunctions}},
  \code{\link{SMRGlobalTermFunctionWeights}},
  \code{\link{SMRMakeDocumentTermMatrix}},
  \code{\link{SMRToBagOfWords}}
}
\concept{Document-term matrix functions}
