% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRMakeDocumentTermMatrix}
\alias{SMRMakeDocumentTermMatrix}
\title{Create document-term matrix from a document collection.}
\usage{
SMRMakeDocumentTermMatrix(documents, ids = NULL, split = "\\\\W",
  applyWordStemming = TRUE, minWordLength = 2)
}
\arguments{
\item{documents}{A list of documents.}

\item{ids}{A list of ID's corresponding to the documents.}

\item{split}{A string pattern to split with.}

\item{applyWordStemming}{Should word stemming be applied or not?}
}
\value{
Sparse matrix
}
\description{
Creates a document-term matrix from a list of documents and list of ID's
}
\details{
It would be nice if this function uses SMRToBagOfWords function defined above.
}
\seealso{
Other Document-term matrix functions: \code{\link{SMRApplyTermWeightFunctions}},
  \code{\link{SMRGlobalTermFunctionWeights}},
  \code{\link{SMRMakeColumnStochastic}},
  \code{\link{SMRToBagOfWords}}
}
\concept{Document-term matrix functions}
