% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRMakeRowStochastic}
\alias{SMRMakeRowStochastic}
\title{Make a matrix row stochastic.}
\usage{
SMRMakeRowStochastic(mat)
}
\arguments{
\item{mat}{A matrix (sparse or dense).}
}
\value{
Sparse matrix
}
\description{
Makes the matrix argument a row stochastic matrix (the sum of each row is 1).
}
\concept{Document-term matrix functions.}
