% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatricesToLongForm}
\alias{SMRMatricesToLongForm}
\title{Sub-matrices to data frame long form conversion}
\usage{
SMRMatricesToLongForm(
  smr,
  tagTypes = NULL,
  removeTagTypePrefixesQ = FALSE,
  sep = ":"
)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagTypes}{A vector tag types (strings) to make the data frame with.
If NULL all tag types are used.}

\item{removeTagTypePrefixesQ}{Should the tag type prefixes be removed from the tags?}

\item{sep}{Separator between tag type prefixes and tags.}
}
\value{
A data frame.
}
\description{
Long form of the data frame.
}
