% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatricesToWideDF}
\alias{SMRMatricesToWideDF}
\title{Sub-matrices to data frame wide form conversion.}
\usage{
SMRMatricesToWideDF(smr, tagTypes = NULL, sep = ", ",
  .progress = "none")
}
\arguments{
\item{smr}{a sparse matrix object}

\item{tagTypes}{the tag types to make the data frame with}

\item{.progress}{Same as \code{.progress} argument for \code{plyr::llply}.}
}
\value{
Data frame
}
\description{
Long form of the data frame
}
