% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatricesToWideForm}
\alias{SMRMatricesToWideForm}
\title{Sub-matrices to data frame wide form conversion}
\usage{
SMRMatricesToWideForm(
  smr,
  tagTypes = NULL,
  sep = ", ",
  removeTagTypePrefixesQ = FALSE,
  tagTypeSep = ":"
)
}
\arguments{
\item{smr}{A sparse matrix object.}

\item{tagTypes}{The tag types to make the data frame with.}

\item{sep}{Separator placed between the tags.}

\item{removeTagTypePrefixesQ}{Should the tag type prefixes be removed from the tags.}

\item{tagTypeSep}{Separator between tag type prefixes and tags.}
}
\value{
A data frame.
}
\description{
Long form of the data frame.
}
