% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatrixDistances}
\alias{SMRMatrixDistances}
\title{Euclidean distances from a vector.}
\usage{
SMRMatrixDistances(smat, vec = colMeans(smat), method = "euclidean")
}
\arguments{
\item{smat}{A sparse matrix.}

\item{vec}{A numeric vector.}

\item{method}{A method for distance computation.
One of "cosine", "euclidean".}
}
\value{
A numeric vector
}
\description{
Computes the Euclidean distances of the rows of a sparse matrix
to a specified vector.
}
\details{
The implementation attempts to keep all computational steps
in sparse matrix structures. 
Note, with "cosine" we mean "cosine distance", not "cosine similarity".
}
