% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatrixEuclideanDistances}
\alias{SMRMatrixEuclideanDistances}
\title{Euclidean distances from a vector.}
\usage{
SMRMatrixEuclideanDistances(smat, vec = colMeans(smat))
}
\arguments{
\item{smat}{A sparse matrix.}

\item{vec}{A numeric vector.}
}
\value{
A numeric vector
}
\description{
Computes the Euclidean distances of the rows of a sparse matrix
to a specified vector.
}
\details{
The implementation attempts to keep all computational steps
in sparse matrix structures.
}
