% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRNormalizeSubMatricesByMaxEntry}
\alias{SMRNormalizeSubMatricesByMaxEntry}
\title{Normalization by maximum.}
\usage{
SMRNormalizeSubMatricesByMaxEntry(smr)
}
\arguments{
\item{smr}{A sparse matrix recommender object (list with named elements).}
}
\value{
Sparse matrix
}
\description{
Makes all sub-matrices to have elements between 0 and 1.
}
