% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRProfileDF}
\alias{SMRProfileDF}
\title{Profile vector calculation.}
\usage{
SMRProfileDF(smr, itemHistory)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{itemHistory}{A data frame with item history with column names c("Rating", "Item").}
}
\value{
A data frame with columns "Score", "Index", "Tag".
}
\description{
Calculate profile from item history.
}
