% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRProfileDFFromVector}
\alias{SMRProfileDFFromVector}
\title{Conversion of a profile vector to a data frame}
\usage{
SMRProfileDFFromVector(smr, pvec)
}
\arguments{
\item{smr}{A sparse matrix recommendation object.}

\item{pvec}{A sparse matrix with one column.}
}
\value{
A data frame with columns "Score", "Index", "Tag".
}
\description{
Convert a data frame corresponding to a profile vector.
}
