% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRProfileDFToVector}
\alias{SMRProfileDFToVector}
\title{Conversion of a profile data frame to a vector.}
\usage{
SMRProfileDFToVector(smr, profileDF, tagType = NULL,
  uniqueColumns = TRUE)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagType}{Tag type over which the vector is made.}

\item{uniqueColumns}{Should the tags in the profile have unique indices in the columns of \code{smr$M}?}

\item{profile}{A data frame with names "Score", "Index", "Tag".}
}
\value{
A sparse matrix with one column.
}
\description{
Return a vector corresponding to a profile data frame.
}
