% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRProfileVector}
\alias{SMRProfileVector}
\title{Profile vector calculation}
\usage{
SMRProfileVector(smr, itemHistory)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{itemHistory}{A data frame with items history with column names \code{ c("Rating", smr$ItemColumnName)}.}
}
\value{
Sparse matrix.
}
\description{
Calculate profile vector from item history.
}
