% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendations}
\alias{SMRRecommendations}
\title{Compute recommendations}
\usage{
SMRRecommendations(
  smr,
  userHistoryItems,
  userRatings,
  nrecs,
  removeHistory = TRUE
)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{userHistoryItems}{The items the user has consumed / purchased.}

\item{userRatings}{Ratings of the history items.}

\item{nrecs}{Number of recommendations to be returned.
If NULL all non-zero score items are returned.}

\item{removeHistory}{Should the history be removed from the recommendations?}
}
\value{
A data frame with columns "Score", "Index", \code{smr$ItemColumnName}.
}
\description{
Recommend items based on a sparse matrix and user history of consumption.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRRecommendationsByProfileDF}()},
\code{\link{SMRRecommendationsByProfileVector}()},
\code{\link{SMRRecommendationsByProfile}()},
\code{\link{SMRRecommendationsDF}()}
}
\concept{Recommendations computation functions}
