% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendationsByProfile}
\alias{SMRRecommendationsByProfile}
\title{Recommendations using two vectors comprising a profile.}
\usage{
SMRRecommendationsByProfile(smr, profileInds, profileRatings, nrecs)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{profileInds}{A vector of metadata indices corresponding to the columns of \code{smr$M}.}

\item{profileRatings}{A vector of ratings of the profile metadata.}

\item{nrecs}{Number of recommendations to be returned.}
}
\value{
A data frame with columns "Score", "Index", "Item".
}
\description{
Recommend items based on a sparse matrix and a specified profile indices and scores.
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRRecommendationsByProfileDF}},
  \code{\link{SMRRecommendationsByProfileVector}},
  \code{\link{SMRRecommendationsDF}},
  \code{\link{SMRRecommendations}}
}
\concept{Recommendations computation functions}
