% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendationsByProfileDF}
\alias{SMRRecommendationsByProfileDF}
\title{Recommendations using a profile data frame}
\usage{
SMRRecommendationsByProfileDF(smr, profile, nrecs)
}
\arguments{
\item{profile}{A data frame of scored tags, profile of a user with column names c( "Score", "Tag" | "Index" )/}

\item{nrecs}{Number of recommendations to be returned,}

\item{smar}{A sparse matrix recommender.}
}
\value{
A data frame with columns "Score", "Index", "Item".
}
\description{
Recommend items based on a sparse matrix and a specified profile.
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRRecommendationsByProfileVector}},
  \code{\link{SMRRecommendationsByProfile}},
  \code{\link{SMRRecommendationsDF}},
  \code{\link{SMRRecommendations}}
}
\concept{Recommendations computation functions}
