% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendationsDF}
\alias{SMRRecommendationsDF}
\title{Compute recommendations using a history data frame.}
\usage{
SMRRecommendationsDF(smr, history, nrecs, removeHistory = TRUE)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{history}{A data frame of rated items with columns "Ratings", <some-item-ID>.}

\item{nrecs}{Number of recommendations to be returned.}

\item{removeHistory}{Should the history be removed from the recommendations?}
}
\value{
A data frame with columns "Score", "Index", \code{smr$ItemColumnName}.
}
\description{
Recommend items based on a sparse matrix recommender and a data frame user history of consumption.
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRRecommendationsByProfileDF}},
  \code{\link{SMRRecommendationsByProfileVector}},
  \code{\link{SMRRecommendationsByProfile}},
  \code{\link{SMRRecommendations}}
}
\concept{Recommendations computation functions}
