% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendationsVectorToDF}
\alias{SMRRecommendationsVectorToDF}
\title{Convert a recommendations vector to a data frame}
\usage{
SMRRecommendationsVectorToDF(rvec, history, nrecs, removeHistory)
}
\arguments{
\item{rvec}{Recommendations vector.}

\item{history}{History of items.}

\item{nrecs}{Number of recommendations to be returned.}

\item{removeHistory}{Should the history be dropped or not?}
}
\value{
Data frame.
}
\description{
Convert a recommendations vector to a data frame.
}
