% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRReorderRecommendations}
\alias{SMRReorderRecommendations}
\title{Recommendations reordering}
\usage{
SMRReorderRecommendations(smr, recs, tagIDs)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{recs}{A data frame recommended items, the second column being row names or row indices.}

\item{tagIDs}{A vector tag ID's of indices with which the recommendations are scored.}
}
\description{
Re-orders a list of recommendations according to
their weighted intersection with a list of tags.
}
\details{
The first column is expected to be of scores.
The original Mathematica package function is named \code{InterfaceUserToLoveFiltered}.
}
\concept{Recommendations functions}
