% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRowBind}
\alias{SMRRowBind}
\title{Row bind two sparse matrix recommenders}
\usage{
SMRRowBind(smr1, smr2)
}
\arguments{
\item{smr1}{The first SMR object.}

\item{smr2}{The second SMR object.}
}
\value{
Sparse matrix recommender.
}
\description{
Returns a sparse matrix recommender object with a matrix
that is obtained by row-binding the matrices of two sparse matrix recommenders.
}
\seealso{
Other SMR modification: \code{\link{SMRRemoveTagTypes}}
}
\concept{SMR modification}
