% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSameTags}
\alias{SMRSameTags}
\title{Same tags.}
\usage{
SMRSameTags(smrTo, smrFrom)
}
\arguments{
\item{smrTo}{The SMR object to be modified.}

\item{smrFrom}{The SMR object the tags of which are imposed.}
}
\value{
Sparse matrix recommender.
}
\description{
Makes the tags of a given sparse matrix recommender object
to be the same as the tags of another one.
}
\details{
The sub-matrices corresponding to each tag type are modified
separately. 
If \code{smrTo} does not have a certain tag type of \code{smrFrom},
then a corresponding empty sub-matrix is created.
}
\seealso{
Other SMR modification: 
\code{\link{SMRRemoveTagTypes}()},
\code{\link{SMRRowBind}()}
}
\concept{SMR modification}
