% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRSparseMatrixCor}
\alias{SMRSparseMatrixCor}
\title{Correlation for a sparse matrix}
\usage{
SMRSparseMatrixCor(x)
}
\arguments{
\item{x}{A sparse matrix}
}
\value{
A matrix
}
\description{
Gives the correlation of the columns of a sparse matrix.
}
\details{
Should produces the same results as \code{cor(as.matrix(x))}.
(Follows the namings of \link{\code{cor}}.)
}
