% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixImportFromDirectory}
\alias{SMRSparseMatrixImportFromDirectory}
\title{Imports sparse matrix from a directory}
\usage{
SMRSparseMatrixImportFromDirectory(
  directoryPath,
  dataNamePrefix = "",
  dataNameInfix = "",
  format = NULL
)
}
\arguments{
\item{directoryPath}{A path to a directory.}

\item{dataNamePrefix}{A string prefix designating the exported SMR object.}

\item{dataNameInfix}{A string infix designating the exported SMR object.}

\item{format}{Export format, one of "CSV", "CSVHarwellBoeing", "feather", or NULL.
If NULL then "CSVHarwellBoeing" is used.}
}
\value{
A sparse matrix
}
\description{
Import a sparse matrix from a specified directory.
}
\details{
If the format "CSVHarwellBoeing" then the sparse matrix is 
imported using \code{\link{readMM}}.
}
