% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixQ}
\alias{SMRSparseMatrixQ}
\title{Is the argument a sparse matrix?}
\usage{
SMRSparseMatrixQ(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
Logical
}
\description{
Gives TRUE if the argument is a sparse matrix object.
}
\details{

}
