% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixRecommenderQ}
\alias{SMRSparseMatrixRecommenderQ}
\title{Is the argument a sparse matrix recommender?}
\usage{
SMRSparseMatrixRecommenderQ(obj)
}
\arguments{
\item{obj}{An object to check.}
}
\value{
Logical
}
\description{
Gives TRUE if the argument is a sparse matrix recommender object.
}
\details{
Is \code{obj} a list and does it have the named elements: \code{c( "M", "M01", "TagTypeRanges", "TagTypes", "ItemColumnName" )}.
}
