% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRSparseMatrixSelectByRow}
\alias{SMRSparseMatrixSelectByRow}
\title{Sparse matrix row-wise element selection}
\usage{
SMRSparseMatrixSelectByRow(smat, selectionCriteria)
}
\arguments{
\item{smat}{A sparse matrix}

\item{selectionCriteria}{Tag selection criteria.
If a positive integer for each row the number of \code{selectionCriteria} top elements are taken.
If a function that function is expected to give a list of booleans for a given list of numerical values
The elements with values that correspond to TRUE are selected.}
}
\description{
Selects elements row-wise using a specified criteria.
}
