% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixToDataFrame}
\alias{SMRSparseMatrixToDataFrame}
\title{Sub-matrix to data frame conversion}
\usage{
SMRSparseMatrixToDataFrame(smr, tagType)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagType}{A tag type (string).}
}
\value{
A data frame.
}
\description{
Makes a data frame of the sparse matrix recommender sub-matrix
that corresponds to a tag type.
}
