% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixToDF}
\alias{SMRSparseMatrixToDF}
\title{Sub-matrix to data frame conversion.}
\usage{
SMRSparseMatrixToDF(smr, tagType)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagType}{A tag type (string).}
}
\value{
Data frame
}
\description{
Makes a data frame of the sparse matrix recommender sub-matrix
that corresponds to a tag type.
}
