% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixToTriplets}
\alias{SMRSparseMatrixToTriplets}
\title{Convert sparse matrix into triplets.}
\usage{
SMRSparseMatrixToTriplets(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\value{
A data frame of triplets
}
\description{
Converts a sparse matrix into a data frame triplets.
}
\seealso{
Other Sparse matrix transformations: \code{\link{SMRImposeColumnIDs}},
  \code{\link{SMRImposeRowIDs}},
  \code{\link{SMRTripletsToSparseMatrix}}
}
\concept{Sparse matrix transformations}
