% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSubMatrix}
\alias{SMRSubMatrix}
\title{Sub-matrix corresponding to a tag type.}
\usage{
SMRSubMatrix(smr, tagType)
}
\arguments{
\item{smr}{A sparse matrix recommender object (list with named elements).}

\item{tagType}{A tag type.}
}
\value{
Sparse matrix
}
\description{
Returns the sub-matrix of the SMR metadata matrix that corresponds to a tag type.
}
