% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSubMatrixOfMatrix}
\alias{SMRSubMatrixOfMatrix}
\title{Sub-matrix corresponding to a tag type.}
\usage{
SMRSubMatrixOfMatrix(M, ranges, tagType)
}
\arguments{
\item{M}{A sparse matrix (in a sparse matrix recommender object).}

\item{ranges}{Column ranges of sub-matrices (in a sparse matrix recommender object).}

\item{tagType}{A tag type.}
}
\description{
Returns the sub-matrix of a matrix that corresponds to a tag type in an SMR object.
}
