% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRTagType}
\alias{SMRTagType}
\title{Tag type correspondence}
\usage{
SMRTagType(smr, tag)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tag}{A type ID (string) or NULL.}
}
\description{
Finds the tag type of a tag.
}
