% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRTagTypeRepresentation}
\alias{SMRTagTypeRepresentation}
\title{SMR's tag types representation}
\usage{
SMRTagTypeRepresentation(recommenders)
}
\arguments{
\item{recommenders}{A list of recommenders.}
}
\value{
A data frame.
}
\description{
Gives a data frame with the tag types representation statistics for a list of recommenders.
}
\details{
If the list of recommenders does not have element names automatic ordinal names are assigned.
}
