% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRToBagOfWords}
\alias{SMRToBagOfWords}
\title{Conversion to bag of words.}
\usage{
SMRToBagOfWords(
  text,
  split = "\\\\W",
  punctuationPattern = "[[:punct:]]",
  stopWords = NULL,
  minWordLength = 2,
  patternToIgnore = NULL
)
}
\arguments{
\item{text}{The text to be converted.}

\item{split}{Character vector containing regular expressions.}

\item{punctuationPattern}{Pattern for the punctuation signs.}

\item{patternToIgnore}{Pattern for words to be ignored.}
}
\description{
Converts a string into a bag of words.
}
\seealso{
Other Document-term matrix functions: 
\code{\link{SMRApplyTermWeightFunctions}()},
\code{\link{SMRGlobalTermFunctionWeights}()},
\code{\link{SMRMakeColumnStochastic}()},
\code{\link{SMRMakeDocumentTermMatrix}()}
}
\concept{Document-term matrix functions}
