% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRToMetadataRecommender}
\alias{SMRToMetadataRecommender}
\title{Convert to a metadata recommender}
\usage{
SMRToMetadataRecommender(
  smr,
  tagTypeTo,
  tagTypes = NULL,
  tagTypeMatrix = NULL,
  normalizerFunc = NULL
)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagTypeTo}{Tag type to make a recommender for.}

\item{tagTypes}{A vector tag types (strings) to make the data frame with. 
If NULL all tag types are used. Passed to \code{\link{SMRMatricesToLongForm}}.}

\item{tagTypeMatrix}{A sparse matrix of item IDs vs tags of \code{tagTypeTo}.
If NULL then \code{SMRSubMatrix(smr, tagTypeTo)} is used.}

\item{normalizerFunc}{An LSI normalizer function.
One of NULL, "None", "Cosine", "Sum", or "Max".
If NULL then it is same as "None". 
See \code{SMRApplyTermWeightFunctions}}
}
\value{
A sparse matrix recommender
}
\description{
Converts the recommender object into a recommender for a 
specified tag type using replacements in the long form representation of the
recommender matrix.
}
\details{

}
