% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRTripletsToSparseMatrix}
\alias{SMRTripletsToSparseMatrix}
\title{Convert a data frame of triplets into a sparse matrix.}
\usage{
SMRTripletsToSparseMatrix(triplets)
}
\arguments{
\item{triplets}{A data frame with three columns.}
}
\value{
A sparse matrix
}
\description{
Turns a data frame of three columns (triplets) into a sparse matrix.
}
\seealso{
Other Sparse matrix transformations: \code{\link{SMRImposeColumnIDs}},
  \code{\link{SMRImposeRowIDs}},
  \code{\link{SMRSparseMatrixToTriplets}}
}
\concept{Sparse matrix transformations}
