% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTermWeightFunctions.R
\name{SMRUnitize}
\alias{SMRUnitize}
\title{Unitize a sparse matrix.}
\usage{
SMRUnitize(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\value{
Sparse matrix
}
\description{
Replaces non-zero entries with 1's.
}
\seealso{
Other Document-term matrix functions: 
\code{\link{SMRApplyTermWeightFunctions}()},
\code{\link{SMRClip}()},
\code{\link{SMRGlobalTermFunctionWeights}()},
\code{\link{SMRMakeColumnStochastic}()},
\code{\link{SMRMakeDocumentTermMatrix}()},
\code{\link{SMRToBagOfWords}()}
}
\concept{Document-term matrix functions}
