% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{TSCorrSMRCreate}
\alias{TSCorrSMRCreate}
\title{Time series recommender creation.}
\usage{
TSCorrSMRCreate(timeSeriesMatrix, smr, classLabelMatrix = NULL,
  correlationMethod = "pearson", smrNRecs = 2000,
  itemIDtoNameRules = NULL)
}
\arguments{
\item{timeSeriesMatrix}{A time series matrix.}

\item{smr}{A sparse matrix recommender.}

\item{classLabelMatrix}{A matrix with class labels; each column is for a different label.}

\item{correlationMethod}{Corellation method, one of: 'dot', 'pearson', 'spearman', 'kendall'.}

\item{smrNRecs}{The number of recommendations for the intermediate candidate finding using the SMR object \code{smr}.}

\item{itemIDtoNameRules}{Item ID's to names rules.}
}
\value{
A list (an S3 object)
}
\description{
Time series recommender creation.
}
\seealso{
Other Creation functions: \code{\link{SMRCreateFromMatrices}},
  \code{\link{SMRCreateItemTagMatrix}},
  \code{\link{SMRCreate}}
}
\concept{Creation functions}
