% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRMakeServerFunction}
\alias{TSCorrSMRMakeServerFunction}
\title{Time series search interface server function}
\usage{
TSCorrSMRMakeServerFunction(tsSMR, tsSearchVectors)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}
}
\value{
Shiny server function.
}
\description{
Time series search interface server function
}
\concept{Time series search interface functions.}
