% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRMakeUI}
\alias{TSCorrSMRMakeUI}
\title{Time series search interface UI.}
\usage{
TSCorrSMRMakeUI(tsSMR, tsSearchVectors)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}
}
\value{
Shiny UI object
}
\description{
Time series search interface UI.
}
\seealso{
Other Time series search interface: \code{\link{TSCorrSMRCreateSearchInterface}},
  \code{\link{TSCorrSMRMakeServerFunction}}
}
\concept{Time series search interface}
