% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{Recommendations.TSCorrSMR}
\alias{Recommendations.TSCorrSMR}
\title{Time series recommendations calculation}
\usage{
\method{Recommendations}{TSCorrSMR}(
  x,
  historyItems,
  historyRatings,
  nrecs,
  removeHistory = TRUE,
  ...
)
}
\value{
A data frame with columns "Score", "Item".
}
\description{
Specialization of Recommendations for TSCorrSMR objects.
}
