% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRApplyTagTypeWeights}
\alias{SMRApplyTagTypeWeights}
\title{Change tag type weights.}
\usage{
SMRApplyTagTypeWeights(smr, weights)
}
\arguments{
\item{smr}{A sparse matrix recommender object (list with named elements).}

\item{weights}{A list/vector of weights to be applied.}
}
\value{
Sparse Matrix
}
\description{
Changes the weights of tag types of a sparse matrix recommender object.
}
\details{
Note tag types should have "significance factors" not "weights".
}
