% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreate}
\alias{SMRCreate}
\title{Creation of a SMR object from a transactions data frame}
\usage{
SMRCreate(dataRows, tagTypes = names(dataRows)[-1],
  itemColumnName = names(dataRows)[1])
}
\arguments{
\item{dataRows}{The transactions data frame.}

\item{tagTypes}{The name of the column containing the categorical tags.}

\item{itemColumnName}{The name of the column containing the unique items.}
}
\value{
SMR object.
}
\description{
Creates a Sparse Matrix Recommender object from transactions data and a list of tag types.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: \code{\link{SMRCreateFromMatrices}},
  \code{\link{SMRCreateFromSpecification}},
  \code{\link{SMRCreateItemTagMatrix}},
  \code{\link{SMREmptySpecification}},
  \code{\link{TSCorrSMRCreate}}
}
\concept{Creation functions}
