% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromMatrices}
\alias{SMRCreateFromMatrices}
\title{Creation of a Sparse Matrix Recommender object with matrix}
\usage{
SMRCreateFromMatrices(matrices, tagTypes = NULL,
  itemColumnName = "Item")
}
\arguments{
\item{matrices}{A list of matrices to be spliced into a metadata matrix.}

\item{tagTypes}{Optional vector of matrix names. If NULL the names \code{matrices} are used.}

\item{itemColumnName}{The column name of recommender items (in data and recommendations).}
}
\value{
SMR object.
}
\description{
Creates a sparse matrix recommender from a list of matrices and a corresponding list of tag types.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: \code{\link{SMRCreateFromSpecification}},
  \code{\link{SMRCreateItemTagMatrix}},
  \code{\link{SMRCreate}}, \code{\link{TSCorrSMRCreate}}
}
\concept{Creation functions}
