% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromMatrices}
\alias{SMRCreateFromMatrices}
\title{Creation of a SMR object with a list of matrices}
\usage{
SMRCreateFromMatrices(
  matrices,
  tagTypes = names(matrices),
  itemColumnName = "Item",
  imposeSameRowNamesQ = TRUE,
  addTagTypesToColumnNamesQ = FALSE,
  sep = ":"
)
}
\arguments{
\item{matrices}{A list of matrices to be spliced into a metadata matrix.}

\item{tagTypes}{Vector of matrix names.}

\item{itemColumnName}{The column name of recommender items (in data and recommendations).}

\item{imposeSameRowNamesQ}{Should the union of the row names be imposed on each matrix?}

\item{addTagTypesToColumnNamesQ}{Should the tag types be added as prefixes 
to the column names of the corresponding sub-matrices?
If NULL it is automatically determined: 
if there is overlap between the tags of different tag types
then becomes TRUE (the tag types are added as prefixes.)}

\item{sep}{Separator for the prefixes of the columns names.}
}
\value{
SMR object.
}
\description{
Creates a Sparse Matrix Recommender object from a list of matrices and a corresponding list of tag types.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: 
\code{\link{SMRCreateFromLongForm}()},
\code{\link{SMRCreateFromSpecification}()},
\code{\link{SMRCreateItemTagMatrix}()},
\code{\link{SMRCreate}()},
\code{\link{SMREmptySpecification}()},
\code{\link{TSCorrSMRCreate}()}
}
\concept{Creation functions}
