% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromSpecification}
\alias{SMRCreateFromSpecification}
\title{Creation of a Sparse Matrix Recommender object from a specification}
\usage{
SMRCreateFromSpecification(data, metaDataSpec, itemCol)
}
\arguments{
\item{data}{A transactions-like data frame.}

\item{metaDataSpec}{A data frame with specifications of which columns of
\code{data} to be used and with what weight functions.}

\item{itemCol}{The name of the column containing the unique items.}
}
\value{
SMR object.
}
\description{
Creates a sparse matrix recommender from transactions-like data
and a meta-data specification.
}
\seealso{
Other Creation functions: \code{\link{SMRCreateFromMatrices}},
  \code{\link{SMRCreateItemTagMatrix}},
  \code{\link{SMRCreate}}, \code{\link{TSCorrSMRCreate}}
}
\concept{Creation functions}
