% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRCreateFromSpecification}
\alias{SMRCreateFromSpecification}
\title{Creation of a Sparse Matrix Recommender object from a specification}
\usage{
SMRCreateFromSpecification(data, metaDataSpec, itemColumnName)
}
\arguments{
\item{data}{A transactions-like data frame.}

\item{metaDataSpec}{A data frame with specifications of which columns of
\code{data} to be used and with what weight functions.}

\item{itemColumnName}{The name of the column containing the unique items.}
}
\value{
An SMR object.
}
\description{
Creates a sparse matrix recommender from transactions-like data
and a meta-data specification.
}
\details{
The specification data frame is expected to have the columns names:
\code{c("ColumnName", "ValueColumnName", "GlobalWeightFunction", "LocalWeightFunction", "NormalizerFunction", "NormalizeByMax")}.
The NA values of \code{"ValueColumnName"} are replaced with \code{"None"}. 
(I.e. do not use \code{"None"} as a "value" column name of \code{data}.)
See \link{\code{SMREmptySpecification}}.
}
\seealso{
Other Creation functions: 
\code{\link{SMRCreateFromLongForm}()},
\code{\link{SMRCreateFromMatrices}()},
\code{\link{SMRCreateItemTagMatrix}()},
\code{\link{SMRCreate}()},
\code{\link{SMREmptySpecification}()},
\code{\link{TSCorrSMRCreate}()}
}
\concept{Creation functions}
