% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRExportToDirectory}
\alias{SMRExportToDirectory}
\title{Exports recommender to a directory}
\usage{
SMRExportToDirectory(smr, directoryPath, dataNameInfix = "")
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{directoryPath}{A path to a directory.}

\item{dataNameInfix}{A string designating the exported SMR object.}
}
\value{
A logical.
}
\description{
Exports the recommender to specified directory.
}
\details{
The sparse matrix recommender \code{smr} is
exported using \code{\link{writeMM}}.
}
