% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRExportToDirectory}
\alias{SMRExportToDirectory}
\title{Exports recommender to a directory}
\usage{
SMRExportToDirectory(
  smr,
  directoryPath,
  dataNamePrefix = "",
  dataNameInfix = "",
  format = NULL,
  digits = NULL
)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{directoryPath}{A path to a directory.}

\item{dataNamePrefix}{A string prefix designating the exported SMR object.}

\item{dataNameInfix}{A string infix designating the exported SMR object.}

\item{format}{Export format, one of "CSV", "CSVHarwellBoeing", "feather", or NULL.
If NULL then "CSVHarwellBoeing" is used.}

\item{digits}{An integer for the number of digits to round to.
See \code{\link{round}}.
If NULL no rounding is done.
Only applies to the file with the name pattern "SMR-M01-from-*".}
}
\value{
A logical.
}
\description{
Exports the recommender to specified directory.
}
\details{
The sparse matrix recommender \code{smr} is
exported using \code{\link{SMRSparseMatrixExportToDirectory}}.
The arguments \code{csvTripletsQ} and \code{digits} are used to specify that
the recommender matrix has to be also exported as a CSV file and the 
weights are rounded (if \code{digits} is numeric.)
}
