% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRImposeRowIDs}
\alias{SMRImposeRowIDs}
\title{Impose row ID's}
\usage{
SMRImposeRowIDs(rowIDs, smat)
}
\arguments{
\item{rowIDs}{An array of row ID's.}

\item{smat}{A matrix with named rows.}
}
\value{
Matrix.
}
\description{
Makes sure that the rows of a matrix are in 1-to-1 correspondence to an array of row ID's
}
\seealso{
Other Sparse matrix transformation functions: 
\code{\link{SMRImposeColumnIDs}()},
\code{\link{SMRSparseMatrixToTriplets}()},
\code{\link{SMRTripletsToSparseMatrix}()}
}
\concept{Sparse matrix transformation functions}
