% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRJoin}
\alias{SMRJoin}
\title{Join two sparse matrix recommender objects}
\usage{
SMRJoin(smr1, smr2, joinType = "same", colnamesPrefix1 = NULL,
  colnamesPrefix2 = NULL)
}
\arguments{
\item{smr1}{The first SMR object.}

\item{smr2}{The second SMR object.}

\item{joinType}{Join type, one of \code{c("inner", "left", "outer", "same")}.}

\item{colnamesPrefix1}{The prefix to be concatenated to the colnames of the first SMR object.}

\item{colnamesPrefix2}{The prefix to be concatenated to the colnames of the second SMR object.}
}
\value{
Sparse matrix recommender.
}
\description{
Joins (column-binds) two SMR objects.
}
\details{
The matrices of the given SMR objects are column bound.
The argument \code{joinType} says how the rows of the corresponding matrices are changed.
For example, if \code{joinType = "outer"} the union of the rownames of SMR matrices is
is row-imposed first. See \code{\link{SMRImposeRowIDs}}.
}
\seealso{
Other SMR modification functions: \code{\link{SMRAnnexSubMatrix}},
  \code{\link{SMRRowBindMatrix}}
}
\concept{SMR modification functions}
