% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatricesToLongDF}
\alias{SMRMatricesToLongDF}
\title{Sub-matrices to data frame long form conversion.}
\usage{
SMRMatricesToLongDF(smr, tagTypes = NULL, .progress = "none")
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagTypes}{A vector tag types (strings) to make the data frame with.}

\item{.progress}{Same as \code{.progress} argument for \code{plyr::llply}.}
}
\value{
Data frame
}
\description{
Long form of the data frame.
}
