% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRMatricesToWideDF}
\alias{SMRMatricesToWideDF}
\title{Sub-matrices to data frame wide form conversion}
\usage{
SMRMatricesToWideDF(smr, tagTypes = NULL, sep = ", ", .progress = "none")
}
\arguments{
\item{smr}{A sparse matrix object.}

\item{tagTypes}{The tag types to make the data frame with.}
}
\value{
A data frame.
}
\description{
Long form of the data frame.
}
