% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRecommendationsByProfileVector}
\alias{SMRRecommendationsByProfileVector}
\title{Recommendations by profile vector}
\usage{
SMRRecommendationsByProfileVector(smr, profileVec, nrecs)
}
\arguments{
\item{profileVec}{A sparse matrix with 1 row (a row from a sparse matrix).}

\item{nrecs}{Number of recommendations to be returned.}

\item{smar}{A sparse matrix recommender.}
}
\value{
A data frame with columns \code{ c("Score", "Index", smr$ItemColumnName)}.
}
\description{
Recommend items based on a sparse matrix and specified profile.
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRRecommendationsByProfileDF}},
  \code{\link{SMRRecommendationsByProfile}},
  \code{\link{SMRRecommendationsDF}},
  \code{\link{SMRRecommendations}}
}
\concept{Recommendations computation functions}
