% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRemoveTagTypes}
\alias{SMRRemoveTagTypes}
\title{Remove tag types from a sparse matrix recommender}
\usage{
SMRRemoveTagTypes(smr, removeTagTypes)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{removeTagTypes}{A list of tag types to be removed from \code{smr}.}
}
\value{
Sparse matrix recommender.
}
\description{
Creates an SMR object from a given SMR object by removing specified tag types.
}
\seealso{
Other SMR modification: \code{\link{SMRRowBind}}
}
\concept{SMR modification}
