% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRRowBind}
\alias{SMRRowBind}
\title{Row bind two sparse matrix recommenders}
\usage{
SMRRowBind(smr1, smr2, type = "union")
}
\arguments{
\item{smr1}{The first SMR object.}

\item{smr2}{The second SMR object.}

\item{type}{A string specifying how the tags for each sub-matrix
are derived. One of "union", "intersection", "left", "right".}
}
\value{
Sparse matrix recommender.
}
\description{
Returns a sparse matrix recommender object with a matrix
that is obtained by row-binding the matrices of two sparse matrix recommenders.
}
\details{
The sub-matrices corresponding to each tag type are row-bound 
separately.
}
\seealso{
Other SMR modification: 
\code{\link{SMRRemoveTagTypes}()},
\code{\link{SMRSameTags}()}
}
\concept{SMR modification}
