% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommender.R
\name{SMRSparseMatrixExportToDirectory}
\alias{SMRSparseMatrixExportToDirectory}
\title{Exports sparse matrix to a directory}
\usage{
SMRSparseMatrixExportToDirectory(
  smat,
  directoryPath,
  dataNamePrefix = "",
  dataNameInfix = "",
  csvTripletsQ = FALSE,
  digits = NULL
)
}
\arguments{
\item{smat}{A sparse matrix.}

\item{directoryPath}{A path to a directory.}

\item{dataNamePrefix}{A string prefix for the export files.}

\item{dataNameInfix}{A string infix for the export files.}

\item{csvTripletsQ}{Should the recommender matrix be also exported to a CSV file.}

\item{digits}{An integer for the number of digits to round to.
See \code{\link{round}}.
If NULL no rounding is done.
Only applies to the file with the name pattern "SMR-M01-from-*.csv".}
}
\value{
A logical.
}
\description{
Exports the recommender to specified directory.
}
\details{
The sparse matrix \code{smat} is
exported using \code{\link{writeMM}}.
The arguments \code{csvTripletsQ} and \code{digits} are used to specify that
the matrix has to be also exported as a CSV file and the 
weights are rounded (if \code{digits} is numeric.)
}
