% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRToMetadataRecommender}
\alias{SMRToMetadataRecommender}
\title{Convert to a metadata recommender}
\usage{
SMRToMetadataRecommender(smr, tagTypeTo, nTopTags = 1, tagTypes = NULL, ...)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagTypeTo}{Tag type to make a recommender for.}

\item{nTopTags}{Number of top tags from \code{tagTypeTo} when making item-tag 
replacements.}

\item{tagTypes}{A vector tag types (strings) to make the data frame with. 
If NULL all tag types are used. Passed to \code{\link{SMRMatricesToLongForm}}.}

\item{...}{Additional arguments for \code{\link{SMRMatricesToLongForm}} or \code{\link{SMRCreateFromMatrices}}.}
}
\value{
A sparse matrix recommender
}
\description{
Converts the recommender object into a recommender for a 
specified tag type.
}
\details{
The following steps are taken.
(1) The long form of the recommender is made.
(2) The items are replaced with the top tags of \code{tagTypeTo}.
(3) A new recommender is created with items that are the tags of \code{tagTypeTo}.
}
