% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRToMetadataRecommender}
\alias{SMRToMetadataRecommender}
\title{Convert to a metadata recommender}
\usage{
SMRToMetadataRecommender(smr, tagType, nTopTags = 1, ...)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tagType}{Tag type to make a recommender for.}

\item{nTopTags}{Number of top tags from \code{tagType} when making item-tag 
replacements.}

\item{...}{Additional arguments for \code{SMRMatricesToLongForm}.}
}
\value{
A sparse matrix recommender
}
\description{
Converts the recommender object into a recommender for a 
specified tag type.
}
\details{
The following steps are taken.
(1) The long form of the recommender is made.
(2) The items are replaced with the top tags of \code{tagType}.
(3) A new recommender is created with items that are the tags of \code{tagType}.
}
