% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRUnitizeSubMatrices}
\alias{SMRUnitizeSubMatrices}
\title{Unitize sub-matrices}
\usage{
SMRUnitizeSubMatrices(smr, tol = 0)
}
\arguments{
\item{smr}{A sparse matrix recommender.}

\item{tol}{A positive number. 
Matrix entries with absolute values below \code{tol} are put to 0.}
}
\value{
A sparse matrix recommender
}
\description{
Unitizes the sub-matrices of a sparse matrix recommender.
}
