% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderExtra.R
\name{SMRUnitizeSubMatrices}
\alias{SMRUnitizeSubMatrices}
\title{Unitize sub-matrices}
\usage{
SMRUnitizeSubMatrices(smr)
}
\arguments{
\item{smr}{A sparse matrix recommender.}
}
\value{
A sparse matrix recommender
}
\description{
Unitizes the sub-matrices of a sparse matrix recommender.
}
