% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesRecommender.R
\name{TSPSRCorrelationNNs}
\alias{TSPSRCorrelationNNs}
\title{Time series recommendations finding.}
\usage{
TSPSRCorrelationNNs(timeSeriesMat, smr, itemIDtoNameRules,
  searchRowID = NULL, searchVector = NULL, nrecs = 12,
  smr.nrecs = 2000, method = "pearson")
}
\arguments{
\item{timeSeriesMat}{A matrix with time siries.}

\item{smr}{An SMR object (e.g. based on time averaged signals).}

\item{itemIDtoNameRules}{Item ID's to names rules.}

\item{searchRowID}{Search row ID that is in \code{rownames(timeSeriesMat) == rownames(smr$M)} or NULL.}

\item{searchVector}{Search vector with dimension \code{ncol(timeSeriesMat) == ncol(smr$M)} or NULL.}

\item{nrecs}{Number of recommendations to return; if NULL all recommendations are returned.}

\item{smr.nrecs}{Number of recommendations for the intermediate candidate finding using the SMR object.}

\item{method}{Correlation method; one of { 'dot', 'pearson', 'spearman', 'kendall' }.}
}
\description{
Finds time series recommendations for specified search row ID and / or search vector ID.
}
\details{
The rownames of \code{timeSeriesMat} are composed of channels and item IDs. An item ID can be in several
different channels. Hence we have row IDs like "NYSE:APPL" or "NASDAQ:APPL".
Note, that this is a package internal function and it is not expected to be used on a regular basis;
see \code{\link{Recommendations.TSCorrSMR}}.
}
